% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binit3}
\alias{binit3}
\title{Do the binning of the cross-correlation}
\usage{
binit3(x1s, y1s, x2s, y2s, nbins, r, do_auto)
}
\arguments{
\item{x1s}{Vector of x coordinates for type 1 cells}

\item{y1s}{Vector of y coordinates for type 1 cells}

\item{x2s}{Vector of x coordinates for type 2 cells}

\item{y2s}{Vector of y coordinates for type 2 cells}

\item{nbins}{Number of bins}

\item{r}{Bin width}

\item{do_auto}{Set to TRUE for auto-correlation}
}
\description{
Do the binning of the cross-correlation
}
