% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRP2files.R
\name{DRP2files}
\alias{DRP2files}
\title{Deregressing estimated breeding values - Two long format file}
\usage{
DRP2files(animalData, parentData, animalCol, sireCol, damCol, parentCol,
  ebvName, r2Name, c = 0.5, h2)
}
\arguments{
\item{animalData}{It is animal data file}

\item{parentData}{It is parents data file}

\item{animalCol}{It is the name of the animal's column}

\item{sireCol}{It is the name of the animal's dam column}

\item{damCol}{It is the name of the animal's dam column}

\item{parentCol}{It the name of the parents' column in the parents data file}

\item{ebvName}{It is the name of the EBV column}

\item{r2Name}{It is the name of the accuracy column}

\item{c}{It is the fraction of genetic variance not explained by markers}

\item{h2}{It the heritability of the trait}
}
\value{
A data frame with deregressed proofs, reliability and weights.
}
\description{
This package is easy to use and can be helpful to calculate deregressed proofs, 
             and their reliabilities and weights.
}
\examples{
## Not to run ##

## Example from Garrick et al., (2009)

animalData=data.frame(ID="Animal", sire="Sire", dam="Dam", EBV=15, r2=0.68)
parentData=data.frame(ID=c("Sire", "Dam"), EBV=c(10, 2), r2=c(0.97, 0.36))

DRP2files(animalData=animalData,
          parentData=parentData,
          animalCol = "ID",
          sireCol   = "sire",
          damCol    = "dam",
          parentCol = "ID",
          ebvName   = "EBV",
          r2Name   = "r2",
          c         = 0.5,
          h2        = 0.25)

## End(Not run)

}
\references{
Garrick, D. J., J. F. Taylor, and R. L. Fernando. 2009. 
            Deregressing estimated breeding values and weighting information 
            for genomic regression analyses. Genet. Sel. Evol. 41:55.
}
