% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideDRP.R

\name{wideDRP}
\alias{wideDRP}
\title{Deregressing estimated breeding values - One wide format file}
\usage{
wideDRP(Data, animalId, sireId, damId, c = 0.5, h2, traitName = NULL,
  animalEBV, sireEBV, damEBV, animalr2, sirer2, damr2)
}
\arguments{
\item{Data}{It is the name of the data file}

\item{animalId}{It is the name of the animal's column}

\item{sireId}{It is the name of the sire's column}

\item{damId}{It is the name of the dam's column}

\item{c}{It is the fraction of genetic variance not explained by markers}

\item{h2}{It the heritability of the trait}

\item{traitName}{It the name of the trait}

\item{animalEBV}{It is the name of the animal's EBV column}

\item{sireEBV}{It is the name of the sire's EBV column}

\item{damEBV}{It is the name of the dam's EBV column}

\item{animalr2}{It is the name of the animal's accuracy column}

\item{sirer2}{It is the name of the sire's accuracy column}

\item{damr2}{It is the name of the dam's accuracy column}
}
\value{
A data frame with deregressed proofs, reliability and weights.
}
\description{
This package is easy to use and can be helpful to calculate deregressed proofs, 
             and their reliabilities and weights.
}
\examples{
## Not to run ##

## Example from Garrick et al., (2009)

Dataset=data.frame(animal="A1000", sire="S10", dam="D100", ebv_anim=15, ebv_sire=10, ebv_dam=2,
                   r2_anim=0.68, r2_sire=0.97, r2_dam=0.36, trait="Trait", c=0.5, h2=0.25)

wideDRP(Data     =  Dataset,
        animalId  = "animal",
        sireId    = "sire",
        damId     = "dam",
        animalEBV = "ebv_anim",
        sireEBV   = "ebv_sire",
        damEBV    = "ebv_dam",
        animalr2  = "r2_anim",
        sirer2    = "r2_sire",
        damr2     = "r2_dam",
        traitName = "trait",
        c         =  0.5,
        h2        =  0.25)

## End(Not run)

}
\references{
Garrick, D. J., J. F. Taylor, and R. L. Fernando. 2009. 
            Deregressing estimated breeding values and weighting information 
            for genomic regression analyses. Genet. Sel. Evol. 41:55.
}
